using promise.all() to syncronize many threads
